android: gradle, updated billing AIDL, restructured
