+++ .gitignore
@@ -32,3 +32,6 @@ Thumbs.db
 .idea
 *.iml
 out
+
+.gradle/
+android/build/
