+++ android/AndroidManifest.xml
@@ -5,5 +5,5 @@
 
 
     <uses-sdk android:minSdkVersion="22"/>
-    <application android:label="@string/app_name"/>
+    <!--<application android:label="@string/app_name"/>-->
 </manifest>
