+++ android/src/com/freshplanet/ane/AirInAppPurchase/ExtensionContext.java
@@ -1,22 +1,18 @@
-//////////////////////////////////////////////////////////////////////////////////////
-//
-//  Copyright 2012 Freshplanet (http://freshplanet.com | opensource@freshplanet.com)
-//  
-//  Licensed under the Apache License, Version 2.0 (the "License");
-//  you may not use this file except in compliance with the License.
-//  You may obtain a copy of the License at
-//  
-//    http://www.apache.org/licenses/LICENSE-2.0
-//  
-//  Unless required by applicable law or agreed to in writing, software
-//  distributed under the License is distributed on an "AS IS" BASIS,
-//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//  See the License for the specific language governing permissions and
-//  limitations under the License.
-//  
-//////////////////////////////////////////////////////////////////////////////////////
-
-package com.freshplanet.inapppurchase;
+/**
+ * Copyright 2017 FreshPlanet
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.freshplanet.ane.AirInAppPurchase;
 
 import java.util.HashMap;
 import java.util.List;
@@ -350,12 +346,15 @@ public class ExtensionContext extends FREContext {
         }
     };
 
+    // --------------------------------------------------------------------------------------//
+    //																						 //
+    // 									 	FREContext SETUP								 //
+    // 																						 //
+    // --------------------------------------------------------------------------------------//
+
     /**
      *
-     * FREContext SETUP
-     *
      */
-
     public void dispose() {
 
         _freActivity = null;
@@ -379,6 +378,10 @@ public class ExtensionContext extends FREContext {
         }
     }
 
+    /**
+     *
+     * @return
+     */
     public Map<String, FREFunction> getFunctions() {
 
         Map<String, FREFunction> functionMap = new HashMap<String, FREFunction>();
