+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Tue Feb 14 13:26:52 EST 2017
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-2.14.1-bin.zip
