+++ settings.gradle
@@ -1,0 +1,9 @@
+include ':android'
+
+if (file('internal/internal-settings.gradle').exists()) {
+  apply from: 'internal/internal-settings.gradle'
+}
+
+if (file('local.gradle').exists()) {
+  apply from: 'local.gradle'
+}
